requirements.txt
Django
django-autocomplete-light
django-crispy-forms
django-filter
django-haystack
django-reversion
django-tables2
djangorestframework
Jinja2
jsonschema
mysqlclient
python-dateutil
requests
Whoosh
xlrd
